+++ sqfentity/lib/sqfentity_connection.dart
@@ -34,8 +34,8 @@ class SqfEntityConnectionMobile implements SqfEntityConnectionBase {
     final List<int> bytes =
         data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes);
     try {
-      final dbpath = await getDatabasesPath();
-      final path = '$dbpath/${connection!.databaseName}';
+      final String databasesPath = connection!.databasePath ?? await getDatabasesPath();
+      final path = '$databasesPath/${connection!.databaseName}';
       if (File(path).existsSync()) {
         await databaseFactory.deleteDatabase(path);
         if (File('$path-wal').existsSync()) {
@@ -55,7 +55,7 @@ class SqfEntityConnectionMobile implements SqfEntityConnectionBase {
     final lock = Lock();
     Database? _db;
     await lock.synchronized(() async {
-      final String databasesPath = await getDatabasesPath();
+      final String databasesPath = connection!.databasePath ??  await getDatabasesPath();
       String path = '';
       path = join(databasesPath, connection!.databaseName);
       final file = File(path);